added logic tests
